app/requirements.txt
# WORKING DEPENDENCIES - All critical packages included
flask==2.3.3
werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0

# Critical dependencies to prevent crashes
markdown==3.5.2
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
beautifulsoup4==4.12.2

# Working dependencies
anthropic==0.49.0
pyyaml==6.0.1
python-docx==1.1.0
nltk==3.8.1
boto3==1.34.34
botocore==1.34.34

# ML dependencies for iFlow matching
scikit-learn==1.3.2
numpy==1.26.0
tabulate==0.9.0
matplotlib==3.8.0
termcolor==2.3.0

# Still avoiding the problematic one
# supabase==2.3.4      # This causes pydantic dependency hell
